pyproject.toml
[project]
name = "payment-processor-rinha-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Juan Lima",email = "juanlimalf@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "duckdb (>=1.3.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
]

[tool.poetry]
# Remove packages configuration to avoid installing current project as package

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"